commands now have dedicated file
